+++ core/src/test/java/org/dkf/jed2k/test/UtilsTest.java
@@ -75,12 +75,12 @@ public class UtilsTest {
     public void testFileCoords() {
         // <size, <pieces, blocks_in_last_piece>>
         LinkedList<Pair<Long, Pair<Integer, Integer>>> template = new LinkedList<Pair<Long, Pair<Integer, Integer>>>();
-        template.add(Pair.make(100l, Pair.make(1, 1)));
+        template.add(Pair.make(100L, Pair.make(1, 1)));
         template.add(Pair.make(Constants.PIECE_SIZE, Pair.make(1, (int)(Constants.PIECE_SIZE/Constants.BLOCK_SIZE))));
-        template.add(Pair.make(Constants.PIECE_SIZE*5+Constants.BLOCK_SIZE*10+100, Pair.make(5+1, (int)(10 + 1))));
+        template.add(Pair.make(Constants.PIECE_SIZE*5+Constants.BLOCK_SIZE*10+100, Pair.make(5+1, 10 + 1)));
 
         for(Pair<Long, Pair<Integer, Integer>> value: template) {
-            int pieces = Utils.divCeil(value.left, (long)Constants.PIECE_SIZE).intValue();
+            int pieces = Utils.divCeil(value.left, Constants.PIECE_SIZE).intValue();
             int blocksPerPiece = (Utils.divCeil(Constants.PIECE_SIZE, Constants.BLOCK_SIZE)).intValue();
             int blocksInLastPiece = Utils.divCeil(value.left % Constants.PIECE_SIZE, Constants.BLOCK_SIZE).intValue();
             if (blocksInLastPiece == 0) blocksInLastPiece = blocksPerPiece;
@@ -133,7 +133,7 @@ public class UtilsTest {
 
     @Test
     public void testLinksGeneration() {
-        assertEquals("ed2k://|file|some_file|100|31D6CFE0D16AE931B73C59D7E0C089C0|/", Utils.formatLink("some_file", 100l, Hash.TERMINAL));
+        assertEquals("ed2k://|file|some_file|100|31D6CFE0D16AE931B73C59D7E0C089C0|/", Utils.formatLink("some_file", 100L, Hash.TERMINAL));
     }
 
     @Test
